refactor: DomainCreator 위치 변경
