+++ src/test/java/kitchenpos/domain/DomainCreator.java
@@ -1,18 +1,8 @@
-package kitchenpos.dao;
+package kitchenpos.domain;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.TableGroup;
 
 public class DomainCreator {
     public static Menu createMenu(String name, Long menuGroupId, BigDecimal price) {
